Switch internal scripts to ts
Also adds a formatting script and runs it against all TS files in the samples
