+++ .github/workflows/ci.yml
@@ -11,4 +11,4 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - uses: actions/setup-node@v4
-    - run: node ./.scripts/validate.js
+    - run: npm run validate
