+++ helloworld-test-sample/src/test/suite/index.ts
@@ -7,7 +7,7 @@ export function run(): Promise<void> {
 	const mocha = new Mocha({
 		ui: 'tdd'
 	});
- 
+
 	const testsRoot = path.resolve(__dirname, '..');
 
 	return new Promise((c, e) => {
