+++ lsp-embedded-request-forwarding/client/src/embeddedSupport.ts
@@ -45,7 +45,7 @@ export function isInsideStyleRegion(
 		}
 		token = scanner.scan();
 	}
-	
+
 	return false;
 }
