+++ lsp-log-streaming-sample/client/src/test/diagnostics.test.ts
@@ -8,34 +8,34 @@ import * as assert from 'assert';
 import { getDocUri, activate } from './helper';
 
 describe('Should get diagnostics', () => {
-  const docUri = getDocUri('diagnostics.txt');
+	const docUri = getDocUri('diagnostics.txt');
 
-  it('Diagnoses uppercase texts', async () => {
-    await testDiagnostics(docUri, [
-      { message: 'ANY is all uppercase.', range: toRange(0, 0, 0, 3), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' },
-      { message: 'ANY is all uppercase.', range: toRange(0, 14, 0, 17), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' },
-      { message: 'OS is all uppercase.', range: toRange(0, 18, 0, 20), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' }
-    ]);
-  });
+	it('Diagnoses uppercase texts', async () => {
+		await testDiagnostics(docUri, [
+			{ message: 'ANY is all uppercase.', range: toRange(0, 0, 0, 3), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' },
+			{ message: 'ANY is all uppercase.', range: toRange(0, 14, 0, 17), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' },
+			{ message: 'OS is all uppercase.', range: toRange(0, 18, 0, 20), severity: vscode.DiagnosticSeverity.Warning, source: 'ex' }
+		]);
+	});
 });
 
 function toRange(sLine: number, sChar: number, eLine: number, eChar: number) {
-  const start = new vscode.Position(sLine, sChar);
-  const end = new vscode.Position(eLine, eChar);
-  return new vscode.Range(start, end);
+	const start = new vscode.Position(sLine, sChar);
+	const end = new vscode.Position(eLine, eChar);
+	return new vscode.Range(start, end);
 }
 
 async function testDiagnostics(docUri: vscode.Uri, expectedDiagnostics: vscode.Diagnostic[]) {
-  await activate(docUri);
+	await activate(docUri);
 
-  const actualDiagnostics = vscode.languages.getDiagnostics(docUri);
+	const actualDiagnostics = vscode.languages.getDiagnostics(docUri);
 
-  assert.equal(actualDiagnostics.length, expectedDiagnostics.length);
+	assert.equal(actualDiagnostics.length, expectedDiagnostics.length);
 
-  expectedDiagnostics.forEach((expectedDiagnostic, i) => {
-    const actualDiagnostic = actualDiagnostics[i];
-    assert.equal(actualDiagnostic.message, expectedDiagnostic.message);
-    assert.deepEqual(actualDiagnostic.range, expectedDiagnostic.range);
-    assert.equal(actualDiagnostic.severity, expectedDiagnostic.severity);
-  });
+	expectedDiagnostics.forEach((expectedDiagnostic, i) => {
+		const actualDiagnostic = actualDiagnostics[i];
+		assert.equal(actualDiagnostic.message, expectedDiagnostic.message);
+		assert.deepEqual(actualDiagnostic.range, expectedDiagnostic.range);
+		assert.equal(actualDiagnostic.severity, expectedDiagnostic.severity);
+	});
 }
