+++ lsp-web-extension-sample/client/src/browserClientMain.ts
@@ -29,7 +29,7 @@ export async function activate(context: ExtensionContext) {
 
 	client = createWorkerLanguageClient(context, clientOptions);
 
-	await  client.start();
+	await client.start();
 	console.log('lsp-web-extension-sample server is ready');
 }
