+++ notifications-sample/src/extension.ts
@@ -22,11 +22,11 @@ export function activate(context: vscode.ExtensionContext) {
 	// Notification with actions
 	const showWarningNotificationWithActions = vscode.commands.registerCommand('notifications-sample.showWarningWithActions', async () => {
 		const selection = await vscode.window.showWarningMessage('Warning Notification With Actions', 'Action 1', 'Action 2', 'Action 3');
-		
+
 		if (selection !== undefined) {
 			vscode.window.showInformationMessage(`You selected: ${selection}`, { modal: true });
 		}
-		
+
 	});
 
 	// Progress notification with option to cancel
