+++ webview-codicons-sample/package.json
@@ -42,7 +42,7 @@
 		"typescript": "^5.6.2",
 		"typescript-eslint": "^8.11.0"
 	},
-  "dependencies": {
+	"dependencies": {
 		"@vscode/codicons": "0.0.20"
-  }
+	}
 }
